requirements.txt
azure-identity==1.16.0
azure-storage-blob==12.19.1
PyNaCl==1.5.0
requests==2.31.0